Add more test and throw error with windsurf+claude

+++ leave-system/src/main/java/com/dornor/leave_system/repository/RequestRepository.java
@@ -3,5 +3,5 @@ package com.dornor.leave_system.repository;
 import com.dornor.leave_system.entity.LeaveRequest;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RequestRepository extends JpaRepository<LeaveRequest, Integer> {
+public interface RequestRepository extends JpaRepository<LeaveRequest, Long> {
 }
