+++ leave-system/src/main/java/com/dornor/leave_system/repository/TypeRepository.java
@@ -3,5 +3,5 @@ package com.dornor.leave_system.repository;
 import com.dornor.leave_system.entity.LeaveTypes;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TypeRepository extends JpaRepository<LeaveTypes, Integer> {
+public interface TypeRepository extends JpaRepository<LeaveTypes, Long> {
 }
